# src/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tags": {
      "get": {
        "description": "Returns all curated tags (e.g. vegetarian, gluten-free), ordered alphabetically by localized name. Falls back to English if requested locale is missing.",
        "operationId": "CatalogController_getTags",
        "parameters": [
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Target locale for tag translations (default: en)",
            "schema": {
              "enum": [
                "it",
                "en",
                "fr"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alphabetically ordered list of localized tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List curated tags",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/ingredients": {
      "get": {
        "description": "Returns curated catalog ingredients matching the given category or text query, ordered alphabetically by localized name.",
        "operationId": "CatalogController_getIngredients",
        "parameters": [
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Target locale for ingredient names (default: en)",
            "schema": {
              "enum": [
                "it",
                "en",
                "fr"
              ],
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter ingredients by culinary category",
            "schema": {
              "enum": [
                "produce",
                "dairy_eggs",
                "meat_poultry",
                "seafood",
                "bakery_grains",
                "pantry_spices",
                "legumes_nuts",
                "beverages_liquids",
                "other"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring search across ingredient slugs and translations",
            "schema": {
              "example": "pomodoro",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alphabetically ordered list of matching curated ingredients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List and search curated ingredients",
        "tags": [
          "Catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "WeCook API",
    "description": "RESTful backend API specification for the WeCook recipe sharing social platform.\n\n### Authentication\nAuthentication is based on secure, HttpOnly JWT cookies:\n- **`accessToken`**: Short-lived (10 minutes) JWT sent automatically by browsers or HTTP clients on protected endpoints.\n- **`refresh_token`**: Long-lived (7 days) JWT restricted to the `/api/auth/refresh` path.\n\n### Media Storage & Static Assets\nUploaded recipe covers, step images, gallery photos, videos, and user avatars are served statically under `/uploads/`.\n\n### Localization & Multilingual Engine\nRecipes and catalogs support Italian (`it`), English (`en`), and French (`fr`), powered by LibreTranslate relational projection.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://app.we-cook.it:8443",
      "description": "Staging / Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT Refresh Token stored in HttpOnly cookie (`refresh_token`) for the refresh endpoint"
      }
    },
    "schemas": {
...
      "TagResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique tag identifier",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Canonical tag slug",
            "example": "vegetarian"
          },
          "name": {
            "type": "string",
            "description": "Localized tag display name",
            "example": "Vegetariano"
          }
        },
        "required": [
          "id",
          "slug",
          "name"
        ]
      },
      "IngredientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique ingredient catalog identifier",
            "example": 12
          },
          "slug": {
            "type": "string",
            "description": "Canonical ingredient slug",
            "example": "parmigiano-reggiano"
          },
          "category": {
            "description": "Ingredient culinary category",
            "example": "dairy_eggs",
            "enum": [
              "produce",
              "dairy_eggs",
              "meat_poultry",
              "seafood",
              "bakery_grains",
              "pantry_spices",
              "legumes_nuts",
              "beverages_liquids",
              "other"
            ],
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "Localized ingredient display name",
            "example": "Parmigiano Reggiano DOP"
          }
        },
        "required": [
          "id",
          "slug",
          "category",
          "name"
        ]
      },
...
    }
  }
}